Use full complex FFT and fix complex gradients
In order to more closely match the results of the PyTorch implementation and
allow swapping of kernels, we are using full FFTs in this implementation. More
importantly, we have also fixed the gradients to the complex kernel parameter,
which were previously incorrect. This fix requires the kernel to be
parameterized as two real valued arrays, denoting the real and imaginary
components of the complex kernel.