Correct normalization layers
For the instance norms, we have fixed incorrect arguments to the GroupNorm
layer which were causing the normalization to not actually implement
InstanceNorm. For FourierNetRGB, we have replaced the incorrect usage of
InstanceNorms with BatchNorms. As a minor change, the MultiplaneFourierNet2D is
now renamed to VmapFourierNet2D in order to match the resulting name of the
class due to the vmap transformation.